Refactored event-list
